Fixed issue with damaging potential entities
The previous code would not allow entities to be damaged if they were not an enemy. That included a faction check, which meant that players could not damage entities at all unless they were in a "Hate" relationship with that entity's faction.

The new code changes this check so that entities can be damaged by _merely potential_ enemies.

There are also new changes to handle multiplayer. If the target, leader, or target's leader are players, it calls the vanilla FriendlyFireCheck code that takes multiplayer settings into account.

It also calls the new code to check if entities can see through blocks that collide with their weapons/ammo (missed when merging by hand).