IsAlly will no longer rely on leaders
The UAIFilterIsAlly code called SCoreUtils.IsAlly, but that only checks if the target has the same leader as yourself. It didn't include the entity itself (nor members of its own faction), so it always failed if the NPC didn't have a leader.

Changed by creating a new method, SCoreUtils.IsFriend, that includes itself, and anything that is not a "potential enemy" according to SCoreUtils.IsEnemyOrPotentialEnemy.

This should fix NPCMOD-SP-0024 in the NPCMod bug list.