Refactored entity targeting methods into EntityTargetingUtilities
Also refactored the methods themselves. Quite a bit of logic went into the method that determines "friendly fire" among NPCs (and players, and leaders), and extra helper methods were added that will be filled in later if/when we get feature requests.

The only behavioral change is that now entities don't damage the player at all with friendly fire. They will not be able to damage the player unless the player damages them first.