Fixed GetFactionRelationship
Hopefully for the last time this time.

Our previous understanding was incorrect. Here's how it actually works:

* If self is a player, it will have a non-null faction, but its relationships with all other factions will be 400 (Neutral), so we can't use their faction to accurately determine relationships.
* If self is an NPC with a faction property (or inherited faction property) in its entity class XML, it will have a non-null, named faction as defined in npcs.xml. However, if npcs.xml does not define a relationship to its own faction - and so far, none do - then its relationship to its own faction will be whatever "*" is in npc.xml.
* If self is an NPC without a faction property in its entity class XML (or inherited faction property), it will have a faction ID of zero. This is whatever faction was defined first in npc.xml, In vanilla this is the "none" faction.

You can see this logic at work in EntityAlive.CopyPropertiesFromEntityClass.

For cases where the entity is an NPC, and thus has a named faction from npc.xml, we should use FactionManager.GetRelationshipValue. That accounts for cases where the other entity shares its faction. It also accurately returns its relationship to the player, since the player faction is not null.

For players, we can use the other entity faction's GetRelationship method, which simply returns what is in the internal Relationships array and doesn't do any logic or checking.