Added method to return default friendly fire faction relationship
This should fix the issue where animals (and other entities) that are neutral to the player, cannot be harmed.

It also introduces a method where we can choose this relationship by target type, as opposed to a single cvar for all targets.

It should also help when using the "IsHuman" tags, since even if an entity has that tag, it doesn't necessarily mean it should be immune from damage if it's in a neutral faction. (Consider companion animals that need to be tamed by "hitting" them with a leash.)