Turret targeting fixes
Reworked the Harmony patches for the auto turret and mini turret.

Neither patch will do anything if the target isn't an NPC. This avoids breaking the vanilla game.

If the target is an NPC, it uses the SCore targeting utilities to determine if it should be targeted, and reproduces the vanilla logic to determine if that target is viewable and can be hit.