Quick fix for NPCMOD-SP-0064
Somewhere in the code, something is trying to get either the "Leader" or "Owner" cvar value, without checking to see if it exists first. (I could not track down where.) When this happens, the cvar is created if it doesn't exist, and set with the value of zero.

When doing a playthrough test of a prefab from the prefab builder, the world includes a test trader. That trader happens to have the entity ID of zero.

As a result, all NPCs in test playthroughs will have that trader as their leader, and won't attack the player.

This fix involves ignoring the value of zero for leader and owner IDs.