Update IngredientsFromContainers.cs
fixed some issues with items not being removed correctly.

there is still an issue when crafting items are in backpack(player inventory) since they don't seem to be in the itemstacks.

if this gets solved i think this feature would be fully usable.

maybe the addition of a flag to lootcontainers would help performance(like the Undead Legacy Broadcast feature).