Revert name change to buffOrderGuardHere; add new buff to match code
NPC Core dialogs use the "buffOrderGuardHere" buff directly, so that buff cannot be renamed. Its behavior should also not change, so it should still set the current order to Stay.

As a solution, I created a new "buffOrderGuard" buff, whose name matches what is in the C# code. This buff is like "buffOrderGuardHere" except it sets the current order to Guard. This buff is what the pathing cube code uses.

I also added comments about this behavior in both the C# code and in buffs.xml.