Revenge targets not set on player allies; fixed bounds issue
The revenge targets were being set on all entities in bounds, including the player hires. This is now fixed, and in addition the revenge targets also will not be set on other players in the party, provided those players are protected from friendly fire.

To detect whether an entity is an ally of the player's party, I created a new `IsAllyOfParty` method in `EntityTargetingUtilities`. I did not change any existing methods so there should be no risk of breaking anything.

I also fixed an issue where the POI's full area was not covered (the `Bounds` constructor shrinks the `size` vector argument in half so only a quarter of the prefab was covered).