Create EntityNPCBandit
This is designed to be a drop-in replacement for `EntityEnemySDX`, but instead of extending from `EntityHuman` (which is also what zombies use), it descends from `EntityBandit` instead.

Since this class isn't already in use, it includes the code to save/load the patrol points and guard data. This was not in `EntityEnemySDX` because it could possibly break existing SCore game saves.