#95: Modify UAISCoreUtils.CanSee use of vanilla "See" caches
By making this change, `CanSee` tries to get the target _directly_ from the positive and negative "See" caches.

Once this change was made, the remainder of the `CanSee` code was now executed, and an issue was revealed.

There is a test to see if the NPC's leader is a player. If so, it determines if _the NPC_ can see _its leader_ (using a stealth check). If not, then it can't see _the target_ either.

This was probably meant to be some kind of guard against NPCs attacking sleepers if those sleepers couldn't see the player, but if so, it was implemented incorrectly.

The current code changes determine if _the target_ can see the NPC's leader, not whether _the NPC_ can see its leader. This accomplishes what I believe was the original intent.