Disable ray hit events when entities can't be damaged
This is a Harmony patch on `ItemActionDynamic.hitTarget` to disable the firing of "onSelf[Primary|Secondary]Action[Ray|Graze]Hit" events if the holding entity can't damage the target entity.

This is the cause of the stun baton "shock" issue with friendly NPCs. The stun baton itself wasn't hitting or doing damage, but the baton was still charging, and when fully charged, it still shocked the friendly NPC.

I also refactored the logic to determine when an entity should use faction targeting. It was already repeated in two places in the code, and I would need to repeat it a third time in my patch.

I did test to make sure that there were no regressions. I tested with a friendly NPC (Baker), and enemy NPC (Harley), and animal (bear), and zombies (mainly Boe). I made sure that I could still shock everyone except the Baker, and that when they attacked each other, they did damage.